De database functies in de repo's gezet.

+++ app/src/main/java/com/a2bfit/Activity/StartTrainingMetSchemaActivity.java
@@ -130,7 +130,7 @@ public class StartTrainingMetSchemaActivity extends AppCompatActivity {
                     int gewicht = Integer.valueOf(gewichtAsString);
 
 
-                    long trainingID = getIntent().getExtras().getInt("trainingID");
+                    long trainingID = getIntent().getExtras().getLong("trainingID");
                     long oefeningID = ((Oefening)lvOefeningenVanSchema.getItemAtPosition(selectedIndex)).getOefeningID();
 
                     TrainingsOefening currentTrainingsOefening = new TrainingsOefening(trainingID, oefeningID, gewicht);
